Moving utility functions for KSP
